Move QWatsonImportMixin to mainwindow.py
This makes more sense to have this there, otherwise, this will become
confusing to have QWatson mixin all over the place.
